+++ docs/html/search/all_b.js
@@ -1,17 +1,17 @@
 var searchData=
 [
-  ['m_108',['m',['../struct_sim_data.html#a0a3c4152d86786884939add509bfa325',1,'SimData']]],
-  ['make_5fall_5funidirectional_5fcrosses_109',['make_all_unidirectional_crosses',['../group__crossers.html#ga3f1b06b7aa35e6a05918e16d1deabeee',1,'make_all_unidirectional_crosses(SimData *d, int from_group, GenOptions g):&#160;sim-operations.c'],['../group__crossers.html#ga3f1b06b7aa35e6a05918e16d1deabeee',1,'make_all_unidirectional_crosses(SimData *d, int from_group, GenOptions g):&#160;sim-operations.c']]],
-  ['make_5fcrosses_5ffrom_5ffile_110',['make_crosses_from_file',['../group__crossers.html#ga3b67d38c13c7a8bef2d0fc91c0365177',1,'make_crosses_from_file(SimData *d, const char *input_file, GenOptions g):&#160;sim-operations.c'],['../group__crossers.html#ga3b67d38c13c7a8bef2d0fc91c0365177',1,'make_crosses_from_file(SimData *d, const char *input_file, GenOptions g):&#160;sim-operations.c']]],
-  ['make_5fdouble_5fcrosses_5ffrom_5ffile_111',['make_double_crosses_from_file',['../group__crossers.html#gaa7c90b4015e4abce750176c6d891cad2',1,'make_double_crosses_from_file(SimData *d, const char *input_file, GenOptions g):&#160;sim-operations.c'],['../group__crossers.html#gaa7c90b4015e4abce750176c6d891cad2',1,'make_double_crosses_from_file(SimData *d, const char *input_file, GenOptions g):&#160;sim-operations.c']]],
-  ['make_5fdoubled_5fhaploids_112',['make_doubled_haploids',['../group__crossers.html#gaf078d6bc1b5d4234a9b960e6132480aa',1,'make_doubled_haploids(SimData *d, int group, GenOptions g):&#160;sim-operations.c'],['../group__crossers.html#gaf078d6bc1b5d4234a9b960e6132480aa',1,'make_doubled_haploids(SimData *d, int group, GenOptions g):&#160;sim-operations.c']]],
-  ['make_5fn_5fcrosses_5ffrom_5ftop_5fm_5fpercent_113',['make_n_crosses_from_top_m_percent',['../group__crossers.html#ga68aca99a41527b019316c8243e6fc7a7',1,'make_n_crosses_from_top_m_percent(SimData *d, int n, int m, int group, GenOptions g):&#160;sim-operations.c'],['../group__crossers.html#ga68aca99a41527b019316c8243e6fc7a7',1,'make_n_crosses_from_top_m_percent(SimData *d, int n, int m, int group, GenOptions g):&#160;sim-operations.c']]],
-  ['map_114',['map',['../struct_sim_data.html#ae1ae6d192937ca2edaba69edf8308711',1,'SimData']]],
-  ['markerblocks_115',['MarkerBlocks',['../struct_marker_blocks.html',1,'']]],
-  ['markerposition_116',['MarkerPosition',['../struct_marker_position.html',1,'']]],
-  ['markers_117',['markers',['../struct_sim_data.html#afaf6ca10aa3f661d7fccc335a5de2af8',1,'SimData']]],
-  ['markers_5fin_5fblock_118',['markers_in_block',['../struct_marker_blocks.html#a9db677d2c96cd8199e634107a1e669eb',1,'MarkerBlocks']]],
-  ['mathematical_20functions_119',['Mathematical functions',['../group__maths.html',1,'']]],
-  ['matrix_120',['matrix',['../struct_decimal_matrix.html#ad53ead42947e6ac7ec961379aa685585',1,'DecimalMatrix']]],
-  ['meiosis_20simulation_20functions_121',['Meiosis Simulation Functions',['../group__meiosis.html',1,'']]]
+  ['m_109',['m',['../struct_sim_data.html#a0a3c4152d86786884939add509bfa325',1,'SimData']]],
+  ['make_5fall_5funidirectional_5fcrosses_110',['make_all_unidirectional_crosses',['../group__crossers.html#ga3f1b06b7aa35e6a05918e16d1deabeee',1,'make_all_unidirectional_crosses(SimData *d, int from_group, GenOptions g):&#160;sim-operations.c'],['../group__crossers.html#ga3f1b06b7aa35e6a05918e16d1deabeee',1,'make_all_unidirectional_crosses(SimData *d, int from_group, GenOptions g):&#160;sim-operations.c']]],
+  ['make_5fcrosses_5ffrom_5ffile_111',['make_crosses_from_file',['../group__crossers.html#ga3b67d38c13c7a8bef2d0fc91c0365177',1,'make_crosses_from_file(SimData *d, const char *input_file, GenOptions g):&#160;sim-operations.c'],['../group__crossers.html#ga3b67d38c13c7a8bef2d0fc91c0365177',1,'make_crosses_from_file(SimData *d, const char *input_file, GenOptions g):&#160;sim-operations.c']]],
+  ['make_5fdouble_5fcrosses_5ffrom_5ffile_112',['make_double_crosses_from_file',['../group__crossers.html#gaa7c90b4015e4abce750176c6d891cad2',1,'make_double_crosses_from_file(SimData *d, const char *input_file, GenOptions g):&#160;sim-operations.c'],['../group__crossers.html#gaa7c90b4015e4abce750176c6d891cad2',1,'make_double_crosses_from_file(SimData *d, const char *input_file, GenOptions g):&#160;sim-operations.c']]],
+  ['make_5fdoubled_5fhaploids_113',['make_doubled_haploids',['../group__crossers.html#gaf078d6bc1b5d4234a9b960e6132480aa',1,'make_doubled_haploids(SimData *d, int group, GenOptions g):&#160;sim-operations.c'],['../group__crossers.html#gaf078d6bc1b5d4234a9b960e6132480aa',1,'make_doubled_haploids(SimData *d, int group, GenOptions g):&#160;sim-operations.c']]],
+  ['make_5fn_5fcrosses_5ffrom_5ftop_5fm_5fpercent_114',['make_n_crosses_from_top_m_percent',['../group__crossers.html#ga68aca99a41527b019316c8243e6fc7a7',1,'make_n_crosses_from_top_m_percent(SimData *d, int n, int m, int group, GenOptions g):&#160;sim-operations.c'],['../group__crossers.html#ga68aca99a41527b019316c8243e6fc7a7',1,'make_n_crosses_from_top_m_percent(SimData *d, int n, int m, int group, GenOptions g):&#160;sim-operations.c']]],
+  ['map_115',['map',['../struct_sim_data.html#ae1ae6d192937ca2edaba69edf8308711',1,'SimData']]],
+  ['markerblocks_116',['MarkerBlocks',['../struct_marker_blocks.html',1,'']]],
+  ['markerposition_117',['MarkerPosition',['../struct_marker_position.html',1,'']]],
+  ['markers_118',['markers',['../struct_sim_data.html#afaf6ca10aa3f661d7fccc335a5de2af8',1,'SimData']]],
+  ['markers_5fin_5fblock_119',['markers_in_block',['../struct_marker_blocks.html#a9db677d2c96cd8199e634107a1e669eb',1,'MarkerBlocks']]],
+  ['mathematical_20functions_120',['Mathematical functions',['../group__maths.html',1,'']]],
+  ['matrix_121',['matrix',['../struct_decimal_matrix.html#ad53ead42947e6ac7ec961379aa685585',1,'DecimalMatrix']]],
+  ['meiosis_20simulation_20functions_122',['Meiosis Simulation Functions',['../group__meiosis.html',1,'']]]
 ];