+++ docs/html/search/all_f.js
@@ -1,12 +1,12 @@
 var searchData=
 [
-  ['rand_5fhalfpoint_137',['RAND_HALFPOINT',['../sim-operations_8c.html#a6db8fbe9028cc259012cbdf075c222b2',1,'RAND_HALFPOINT():&#160;sim-operations.c'],['../group__maths.html#ga6db8fbe9028cc259012cbdf075c222b2',1,'RAND_HALFPOINT():&#160;sim-operations.h']]],
-  ['randb_138',['randb',['../group__maths.html#gad4e0f7934e82157e3115f8df7aec01f0',1,'randb(void):&#160;sim-operations.c'],['../group__maths.html#gad4e0f7934e82157e3115f8df7aec01f0',1,'randb(void):&#160;sim-operations.c']]],
-  ['randlim_139',['randlim',['../group__maths.html#ga89e066191f12e056cebb8d2cd74a2a29',1,'randlim(int limit):&#160;sim-operations.c'],['../group__maths.html#ga89e066191f12e056cebb8d2cd74a2a29',1,'randlim(int limit):&#160;sim-operations.c']]],
-  ['randn_140',['randn',['../group__maths.html#ga3d910624bfa7ac1bfacf706c3357fc5d',1,'randn(void):&#160;sim-operations.c'],['../group__maths.html#ga3d910624bfa7ac1bfacf706c3357fc5d',1,'randn(void):&#160;sim-operations.c']]],
-  ['randpoi_141',['randpoi',['../group__maths.html#ga7d9dbe538d6fb096af14c3163c98e800',1,'randpoi(double lambda):&#160;sim-operations.c'],['../group__maths.html#ga7d9dbe538d6fb096af14c3163c98e800',1,'randpoi(double lambda):&#160;sim-operations.c']]],
-  ['read_5fblock_5ffile_142',['read_block_file',['../group__calculators.html#ga463082fc380a6815f07bd06401927bc0',1,'read_block_file(SimData *d, const char *block_file):&#160;sim-operations.c'],['../group__calculators.html#ga463082fc380a6815f07bd06401927bc0',1,'read_block_file(SimData *d, const char *block_file):&#160;sim-operations.c']]],
-  ['readme_2emd_143',['README.md',['../_r_e_a_d_m_e_8md.html',1,'']]],
-  ['recombination_20calculators_144',['Recombination Calculators',['../group__recomb.html',1,'']]],
-  ['rows_145',['rows',['../struct_decimal_matrix.html#a7a80ff4671b6ea3fba3a9ecd9e66a028',1,'DecimalMatrix']]]
+  ['rand_5fhalfpoint_138',['RAND_HALFPOINT',['../sim-operations_8c.html#a6db8fbe9028cc259012cbdf075c222b2',1,'RAND_HALFPOINT():&#160;sim-operations.c'],['../group__maths.html#ga6db8fbe9028cc259012cbdf075c222b2',1,'RAND_HALFPOINT():&#160;sim-operations.h']]],
+  ['randb_139',['randb',['../group__maths.html#gad4e0f7934e82157e3115f8df7aec01f0',1,'randb(void):&#160;sim-operations.c'],['../group__maths.html#gad4e0f7934e82157e3115f8df7aec01f0',1,'randb(void):&#160;sim-operations.c']]],
+  ['randlim_140',['randlim',['../group__maths.html#ga89e066191f12e056cebb8d2cd74a2a29',1,'randlim(int limit):&#160;sim-operations.c'],['../group__maths.html#ga89e066191f12e056cebb8d2cd74a2a29',1,'randlim(int limit):&#160;sim-operations.c']]],
+  ['randn_141',['randn',['../group__maths.html#ga3d910624bfa7ac1bfacf706c3357fc5d',1,'randn(void):&#160;sim-operations.c'],['../group__maths.html#ga3d910624bfa7ac1bfacf706c3357fc5d',1,'randn(void):&#160;sim-operations.c']]],
+  ['randpoi_142',['randpoi',['../group__maths.html#ga7d9dbe538d6fb096af14c3163c98e800',1,'randpoi(double lambda):&#160;sim-operations.c'],['../group__maths.html#ga7d9dbe538d6fb096af14c3163c98e800',1,'randpoi(double lambda):&#160;sim-operations.c']]],
+  ['read_5fblock_5ffile_143',['read_block_file',['../group__calculators.html#ga463082fc380a6815f07bd06401927bc0',1,'read_block_file(SimData *d, const char *block_file):&#160;sim-operations.c'],['../group__calculators.html#ga463082fc380a6815f07bd06401927bc0',1,'read_block_file(SimData *d, const char *block_file):&#160;sim-operations.c']]],
+  ['readme_2emd_144',['README.md',['../_r_e_a_d_m_e_8md.html',1,'']]],
+  ['recombination_20calculators_145',['Recombination Calculators',['../group__recomb.html',1,'']]],
+  ['rows_146',['rows',['../struct_decimal_matrix.html#a7a80ff4671b6ea3fba3a9ecd9e66a028',1,'DecimalMatrix']]]
 ];