+++ docs/html/search/classes_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['allelematrix_196',['AlleleMatrix',['../struct_allele_matrix.html',1,'']]]
+  ['allelematrix_197',['AlleleMatrix',['../struct_allele_matrix.html',1,'']]]
 ];