+++ docs/html/search/classes_3.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['geneticmap_199',['GeneticMap',['../struct_genetic_map.html',1,'']]],
-  ['genoptions_200',['GenOptions',['../struct_gen_options.html',1,'']]]
+  ['geneticmap_200',['GeneticMap',['../struct_genetic_map.html',1,'']]],
+  ['genoptions_201',['GenOptions',['../struct_gen_options.html',1,'']]]
 ];