+++ docs/html/search/classes_6.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['tablesize_204',['TableSize',['../struct_table_size.html',1,'']]]
+  ['tablesize_205',['TableSize',['../struct_table_size.html',1,'']]]
 ];