+++ docs/html/search/defines_2.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['name_5flength_375',['NAME_LENGTH',['../sim-operations_8h.html#af71324c57f05ff9e24bd384925dd6b17',1,'sim-operations.h']]]
+  ['name_5flength_377',['NAME_LENGTH',['../sim-operations_8h.html#af71324c57f05ff9e24bd384925dd6b17',1,'sim-operations.h']]]
 ];