+++ docs/html/search/functions_5.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['has_5fsame_5falleles_278',['has_same_alleles',['../group__recomb.html#gaa71cdbced90506820ef61bd842ecfa40',1,'sim-operations.h']]],
-  ['has_5fsame_5falleles_5fwindow_279',['has_same_alleles_window',['../group__recomb.html#gab88223532880a22dfc966be4cdb4107b',1,'sim-operations.h']]]
+  ['has_5fsame_5falleles_280',['has_same_alleles',['../group__recomb.html#gaa71cdbced90506820ef61bd842ecfa40',1,'sim-operations.h']]],
+  ['has_5fsame_5falleles_5fwindow_281',['has_same_alleles_window',['../group__recomb.html#gab88223532880a22dfc966be4cdb4107b',1,'sim-operations.h']]]
 ];