+++ docs/html/search/functions_8.js
@@ -1,8 +1,8 @@
 var searchData=
 [
-  ['randb_293',['randb',['../group__maths.html#gad4e0f7934e82157e3115f8df7aec01f0',1,'randb(void):&#160;sim-operations.c'],['../group__maths.html#gad4e0f7934e82157e3115f8df7aec01f0',1,'randb(void):&#160;sim-operations.c']]],
-  ['randlim_294',['randlim',['../group__maths.html#ga89e066191f12e056cebb8d2cd74a2a29',1,'randlim(int limit):&#160;sim-operations.c'],['../group__maths.html#ga89e066191f12e056cebb8d2cd74a2a29',1,'randlim(int limit):&#160;sim-operations.c']]],
-  ['randn_295',['randn',['../group__maths.html#ga3d910624bfa7ac1bfacf706c3357fc5d',1,'randn(void):&#160;sim-operations.c'],['../group__maths.html#ga3d910624bfa7ac1bfacf706c3357fc5d',1,'randn(void):&#160;sim-operations.c']]],
-  ['randpoi_296',['randpoi',['../group__maths.html#ga7d9dbe538d6fb096af14c3163c98e800',1,'randpoi(double lambda):&#160;sim-operations.c'],['../group__maths.html#ga7d9dbe538d6fb096af14c3163c98e800',1,'randpoi(double lambda):&#160;sim-operations.c']]],
-  ['read_5fblock_5ffile_297',['read_block_file',['../group__calculators.html#ga463082fc380a6815f07bd06401927bc0',1,'read_block_file(SimData *d, const char *block_file):&#160;sim-operations.c'],['../group__calculators.html#ga463082fc380a6815f07bd06401927bc0',1,'read_block_file(SimData *d, const char *block_file):&#160;sim-operations.c']]]
+  ['randb_295',['randb',['../group__maths.html#gad4e0f7934e82157e3115f8df7aec01f0',1,'randb(void):&#160;sim-operations.c'],['../group__maths.html#gad4e0f7934e82157e3115f8df7aec01f0',1,'randb(void):&#160;sim-operations.c']]],
+  ['randlim_296',['randlim',['../group__maths.html#ga89e066191f12e056cebb8d2cd74a2a29',1,'randlim(int limit):&#160;sim-operations.c'],['../group__maths.html#ga89e066191f12e056cebb8d2cd74a2a29',1,'randlim(int limit):&#160;sim-operations.c']]],
+  ['randn_297',['randn',['../group__maths.html#ga3d910624bfa7ac1bfacf706c3357fc5d',1,'randn(void):&#160;sim-operations.c'],['../group__maths.html#ga3d910624bfa7ac1bfacf706c3357fc5d',1,'randn(void):&#160;sim-operations.c']]],
+  ['randpoi_298',['randpoi',['../group__maths.html#ga7d9dbe538d6fb096af14c3163c98e800',1,'randpoi(double lambda):&#160;sim-operations.c'],['../group__maths.html#ga7d9dbe538d6fb096af14c3163c98e800',1,'randpoi(double lambda):&#160;sim-operations.c']]],
+  ['read_5fblock_5ffile_299',['read_block_file',['../group__calculators.html#ga463082fc380a6815f07bd06401927bc0',1,'read_block_file(SimData *d, const char *block_file):&#160;sim-operations.c'],['../group__calculators.html#ga463082fc380a6815f07bd06401927bc0',1,'read_block_file(SimData *d, const char *block_file):&#160;sim-operations.c']]]
 ];