+++ docs/html/search/groups_0.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['breeding_20value_20and_20allele_20count_20calculators_378',['Breeding Value and Allele Count Calculators',['../group__calculators.html',1,'']]]
+  ['breeding_20value_20and_20allele_20count_20calculators_380',['Breeding Value and Allele Count Calculators',['../group__calculators.html',1,'']]]
 ];