+++ docs/html/search/groups_2.js
@@ -1,6 +1,6 @@
 var searchData=
 [
-  ['deletor_20functions_380',['Deletor Functions',['../group__deletors.html',1,'']]],
-  ['data_20access_20and_20search_20functions_381',['Data Access and Search Functions',['../group__getters.html',1,'']]],
-  ['data_20structures_382',['Data Structures',['../group__structs.html',1,'']]]
+  ['deletor_20functions_382',['Deletor Functions',['../group__deletors.html',1,'']]],
+  ['data_20access_20and_20search_20functions_383',['Data Access and Search Functions',['../group__getters.html',1,'']]],
+  ['data_20structures_384',['Data Structures',['../group__structs.html',1,'']]]
 ];